requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
click==8.1.8
colorama==0.4.6
Django==5.1.5
djangorestframework==3.15.2
drf-spectacular==0.28.0
fastapi==0.115.7
h11==0.14.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
uvicorn==0.34.0
drf-spectacular
